+++ stream-chat-android-client/src/release/java/io.getstream.chat.android.client.di/ChatModule.kt
@@ -2,5 +2,17 @@ package io.getstream.chat.android.client.di
 
 import android.content.Context
 import io.getstream.chat.android.client.api.ChatClientConfig
+import io.getstream.chat.android.client.notifications.handler.ChatNotificationHandler
+import io.getstream.chat.android.client.token.TokenManager
+import io.getstream.chat.android.client.uploader.FileUploader
 
-internal class ChatModule(appContext: Context, config: ChatClientConfig) : BaseChatModule(appContext, config)
+/**
+ * Release variant of [BaseChatModule].
+ */
+internal class ChatModule(
+    appContext: Context,
+    config: ChatClientConfig,
+    notificationsHandler: ChatNotificationHandler,
+    uploader: FileUploader?,
+    tokenManager: TokenManager,
+) : BaseChatModule(appContext, config, notificationsHandler, uploader, tokenManager)