+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/viewmodel/messages/MessageListViewModelBinding.kt
@@ -42,4 +42,5 @@ public fun MessageListViewModel.bindView(view: MessageListView, lifecycleOwner: 
             }
         }
     }
+    loadMoreLiveData.observe(lifecycleOwner, view::setLoadingMore)
 }