+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Allow setting custom empty state views
 - Add loading view to message list view
 - Allow setting custom loading view
+- Add load more threshold for `MessageListView` and `streamLoadMoreThreshold` attribute
 
 ## stream-chat-android-client
 - Fix guest user authentication